EES-2832 - removed CaptureCancellationToken and AddCapturedCancellation attributes, in favour of passing CancellationToken explicitly
