+++ src/GovUk.Education.ExploreEducationStatistics.Data.Api/Startup.cs
@@ -184,8 +184,8 @@ namespace GovUk.Education.ExploreEducationStatistics.Data.Api
             // Enable caching and register any caching services.
             CacheAspect.Enabled = true;
             BlobCacheAttribute.AddService("default", app.ApplicationServices.GetService<IBlobCacheService>());
-            // Enable cancellation aspects.
-            CancellationAspects.Enabled = true;
+            // Enable cancellation aspects and register request timeout configuration.
+            AddTimeoutCancellationAspect.Enabled = true;
             AddTimeoutCancellationAttribute.SetTimeoutConfiguration(Configuration.GetSection("RequestTimeouts"));
 
             UpdateDatabase(app);
