+++ src/GovUk.Education.ExploreEducationStatistics.Data.Services/Interfaces/ITableBuilderService.cs
@@ -1,5 +1,6 @@
 #nullable enable
 using System;
+using System.Threading;
 using System.Threading.Tasks;
 using GovUk.Education.ExploreEducationStatistics.Common.Model;
 using GovUk.Education.ExploreEducationStatistics.Common.Model.Data.Query;
@@ -10,10 +11,13 @@ namespace GovUk.Education.ExploreEducationStatistics.Data.Services.Interfaces
 {
     public interface ITableBuilderService
     {
-        Task<Either<ActionResult, TableBuilderResultViewModel>> Query(ObservationQueryContext queryContext);
+        Task<Either<ActionResult, TableBuilderResultViewModel>> Query(
+            ObservationQueryContext queryContext,
+            CancellationToken cancellationToken = default);
 
         Task<Either<ActionResult, TableBuilderResultViewModel>> Query(
             Guid releaseId,
-            ObservationQueryContext queryContext);
+            ObservationQueryContext queryContext,
+            CancellationToken cancellationToken = default);
     }
 }
